+++ RabbitMqHttpApiClient/Models/Common/MessageStatsModel/IAck.cs
@@ -5,7 +5,7 @@ namespace RabbitMqHttpApiClient.Models.Common.MessageStatsModel
     public interface IAck
     {
         [JsonProperty("ack")]
-        int Ack { get; set; }
+        long Ack { get; set; }
 
         [JsonProperty("ack_details")]
         RateDetails AckDetails { get; set; }
