+++ RabbitMqHttpApiClient/Models/Common/MessageStatsModel/IDeliverGet.cs
@@ -5,7 +5,7 @@ namespace RabbitMqHttpApiClient.Models.Common.MessageStatsModel
     public interface IDeliverGet
     {
         [JsonProperty("deliver_get")]
-        int DeliverGet { get; set; }
+        long DeliverGet { get; set; }
 
         [JsonProperty("deliver_get_details")]
         RateDetails DeliverGetDetails { get; set; }
