+++ RabbitMqHttpApiClient/Models/Common/MessageStatsModel/IDeliverNoAck.cs
@@ -5,7 +5,7 @@ namespace RabbitMqHttpApiClient.Models.Common.MessageStatsModel
     public interface IDeliverNoAck
     {
         [JsonProperty("deliver_no_ack")]
-        int DeliverNoAck { get; set; }
+        long DeliverNoAck { get; set; }
 
         [JsonProperty("deliver_no_ack_details")]
         RateDetails DeliverNoAckDetails { get; set; }
