+++ RabbitMqHttpApiClient/Models/Common/MessageStatsModel/IPublishInfo.cs
@@ -5,19 +5,19 @@ namespace RabbitMqHttpApiClient.Models.Common.MessageStatsModel
     public interface IPublishInfo
     {
         [JsonProperty("publish")]
-        int Publish { get; set; }
+        long Publish { get; set; }
 
         [JsonProperty("publish_details")]
         RateDetails PublishDetails { get; set; }
 
         [JsonProperty("publish_in")]
-        int PublishIn { get; set; }
+        long PublishIn { get; set; }
 
         [JsonProperty("publish_in_details")]
         RateDetails PublishInDetails { get; set; }
 
         [JsonProperty("publish_out")]
-        int PublishOut { get; set; }
+        long PublishOut { get; set; }
 
         [JsonProperty("publish_out_details")]
         RateDetails PublishOutDetails { get; set; }
